refactor(app/renderer/components): move tests into dedicated folder with component
