+++ app/renderer/src/components/query-field/index.js
@@ -1,0 +1,1 @@
+export { default } from './query-field';
