+++ app/renderer/src/components/result-item/index.js
@@ -1,0 +1,1 @@
+export { default } from './result-item';
