+++ app/renderer/src/components/result-item/result-item.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import ResultItem from '../../../../../app/renderer/src/components/ResultItem';
+import ResultItem from './result-item';
 
 describe('<ResultItem /> component', () => {
   let theme = {};
