+++ app/renderer/src/containers/IconContainer.js
@@ -1,7 +1,7 @@
 import { ipcRenderer } from 'electron';
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
-import Icon from '../components/Icon';
+import Icon from '../components/icon';
 import { IPC_FETCH_ICON, IPC_RETRIEVE_ICON } from '../../../ipc';
 
 export default class IconContainer extends Component {
