+++ app/renderer/src/containers/ResultItemContainer.js
@@ -4,7 +4,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import * as actionCreators from '../actions/creators';
-import ResultItem from '../components/ResultItem';
+import ResultItem from '../components/result-item';
 import { ThemeSchema, ResultItemSchema } from '../schema';
 import { IPC_EXECUTE_ITEM } from '../../../ipc';
